Module amazon_sp_reports::apis::reports_api

Enums

CancelReportError

struct for typed errors of method cancel_report

CancelReportScheduleError

struct for typed errors of method cancel_report_schedule

CreateReportError

struct for typed errors of method create_report

CreateReportScheduleError

struct for typed errors of method create_report_schedule

GetReportDocumentError

struct for typed errors of method get_report_document

GetReportError

struct for typed errors of method get_report

GetReportScheduleError

struct for typed errors of method get_report_schedule

GetReportSchedulesError

struct for typed errors of method get_report_schedules

GetReportsError

struct for typed errors of method get_reports

Functions

cancel_report

Cancels the report that you specify. Only reports with processingStatus=IN_QUEUE can be cancelled. Cancelled reports are returned in subsequent calls to the getReport and getReports operations. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 0.0222 | 10 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

cancel_report_schedule

Cancels the report schedule that you specify. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 0.0222 | 10 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

create_report

Creates a report. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 0.0167 | 15 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

create_report_schedule

Creates a report schedule. If a report schedule with the same report type and marketplace IDs already exists, it will be cancelled and replaced with this one. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 0.0222 | 10 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_report

Returns report details (including the reportDocumentId, if available) for the report that you specify. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 2.0 | 15 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_report_document

Returns the information required for retrieving a report document’s contents. This includes a presigned URL for the report document as well as the information required to decrypt the document’s contents. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 0.0167 | 15 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_report_schedule

Returns report schedule details for the report schedule that you specify. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 0.0222 | 10 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_report_schedules

Returns report schedule details that match the filters that you specify. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 0.0222 | 10 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_reports

Returns report details for the reports that match the filters that you specify. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 0.0222 | 10 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.